requirements.txt
streamlit  
pandas
seaborn 
matplotlib
sklearn.linear_model  
sklearn.model_selection  